Обновление "ИЮЛЬ-31"
Изменения по сравнению с предыдущим коммитом:
- Обновлён бэкап базы данных, исправлены ошибки и недочёты в
конфигурации экспорта
- Добавлена ссылка на административный раздел олимпиад в меню
- Исправлены ошибки, допущенные в предыдущих коммитах
- Мы есть Sirkadirov!
- На этом пока что всё!

+++ services/olymp/user/problems.php
@@ -120,7 +120,7 @@
 		$query->free();
 		unset($query);
 		
-		return (int)$result;
+		return (float)$result;
 	}
 	
 	/////////////////////////////////////
